Collapse ChannelQObject class into QtInProcessChannel
